Test: add test for goroutine
This test will first add 1000 numbers into a tree and then delete some
from it.

All the adding and deleting are in goroutine so that the delete action
will not always succeed for example `delete(400)` but `add(400)` has not
been executed yet.

Finally, we'll add back all the deleted nodes to see if there is
anything wrong.

Signed-off-by: Hu Keping <hukeping@huawei.com>